Fix stroke and fill issue, verified with tiger.svg on laptop

+++ libsvgtiny/src/svgtiny_internal.h
@@ -40,7 +40,7 @@ struct svgtiny_parse_state {
 	/* paint attributes */
 	svgtiny_colour fill;
 	svgtiny_colour stroke;
-	int stroke_width;
+	float stroke_width;
 
 	/* gradients */
 	unsigned int linear_gradient_stop_count;
